pyproject.toml
[tool.poetry]
name = "bitly-fastapi"
version = "0.1.0"
description = ""
authors = ["Nikita Driuma <nikita.dryma@gmail.com>"]
readme = "README.md"
packages = [{include = "bitly_fastapi"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.1"
uvicorn = "^0.19.0"
pyjwt = "^2.6.0"
psycopg2 = "^2.9.5"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
pytest = "^7.2.0"
bcrypt = "^4.0.1"
python-decouple = "^3.6"
requests = "^2.28.1"
pytest-asyncio = "^0.20.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
